Move tests to tests/TestCases
