+++ tests/TestCases/Di/ContainerTest.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * AVOLUTIONS
+ *
+ * Just another open source PHP framework.
+ *
+ * @copyright   Copyright (c) 2019 - 2021 AVOLUTIONS
+ * @license     MIT License (https://avolutions.org/license)
+ * @link        https://avolutions.org
+ */
+
+namespace Avolutions\Test\TestCases\Di;
+
+use PHPUnit\Framework\TestCase;
+
+class ContainerTest extends TestCase
+{
+
+}
